Add files via upload
Python (v3) code for a client program and establish successful communication with the server, which runs on the IP 160.75.154.126. Client program should contain two distinct phases in order to execute the communication successfully. First, it should authenticate itself to the server and then server is going to send a series of questions in the context of “Guess the Word” game and client program should provide the sufficient interface to a user to play the game